+++ crates/controller/src/api/v1/auth/mod.rs
@@ -13,10 +13,11 @@ use crate::settings::SharedSettingsActix;
 use actix_web::web::{Data, Json};
 use actix_web::{get, post};
 use core::mem::take;
-use database::Db;
+use database::{Db, OptionalExt};
 use db_storage::events::EventId;
 use db_storage::groups::{get_or_create_groups_by_name, Group, GroupName};
 use db_storage::rooms::RoomId;
+use db_storage::tariffs::{ExternalTariffId, Tariff};
 use db_storage::tenants::{get_or_create_tenant_by_oidc_id, OidcTenantId, TenantId};
 use db_storage::users::User;
 use kustos::prelude::PoliciesBuilder;
@@ -79,6 +80,15 @@ pub async fn login(
         let settings = settings.load_full();
         let mut conn = db.get_conn()?;
 
+        let tariff =
+            Tariff::get_by_external_id(&mut conn, &ExternalTariffId::from(info.tariff_id.clone()))
+                .optional()?
+                .ok_or_else(|| {
+                    ApiError::internal()
+                        .with_code("invalid_tariff_id")
+                        .with_message("JWT contained unknown tariff_id")
+                })?;
+
         let tenant = get_or_create_tenant_by_oidc_id(
             &mut conn,
             &OidcTenantId::from(info.tenant_id.clone()),
@@ -97,11 +107,11 @@ pub async fn login(
         let login_result = match user {
             Some(user) => {
                 // Found a matching user, update its attributes, tenancy and groups
-                update_user::update_user(&settings, &mut conn, user, info, groups)?
+                update_user::update_user(&settings, &mut conn, user, info, groups, tariff)?
             }
             None => {
                 // No matching user, create a new one with inside the given tenants and groups
-                create_user::create_user(&settings, &mut conn, info, tenant, groups)?
+                create_user::create_user(&settings, &mut conn, info, tenant, groups, tariff)?
             }
         };