+++ crates/db-storage/src/utils.rs
@@ -3,6 +3,13 @@
 // SPDX-License-Identifier: EUPL-1.2
 
 use crate::users::UserId;
+use diesel::deserialize::FromSql;
+use diesel::pg::Pg;
+use diesel::serialize::{IsNull, ToSql};
+use diesel::sql_types;
+use serde::{Deserialize, Serialize};
+use std::fmt::Debug;
+use std::io::Write;
 
 /// Trait for models that have user-ids attached to them like created_by/updated_by fields
 ///
@@ -22,5 +29,34 @@ where
         for t in self {
             t.populate(dst);
         }
+    }
+}
+
+/// JSONB Wrapper for any type implementing the serde `Serialize` or `Deserialize` trait
+#[derive(Debug, Clone, Default, FromSqlRow, AsExpression)]
+#[diesel(sql_type = sql_types::Jsonb)]
+pub struct Jsonb<T>(pub T);
+
+impl<T: for<'de> Deserialize<'de>> FromSql<sql_types::Jsonb, Pg> for Jsonb<T> {
+    fn from_sql(value: diesel::backend::RawValue<'_, Pg>) -> diesel::deserialize::Result<Self> {
+        let bytes = value.as_bytes();
+        if bytes[0] != 1 {
+            return Err("Unsupported JSONB encoding version".into());
+        }
+        serde_json::from_slice(&bytes[1..])
+            .map(Self)
+            .map_err(|_| "Invalid Json".into())
+    }
+}
+
+impl<T: Serialize + Debug> ToSql<sql_types::Jsonb, Pg> for Jsonb<T> {
+    fn to_sql<'b>(
+        &'b self,
+        out: &mut diesel::serialize::Output<'b, '_, Pg>,
+    ) -> diesel::serialize::Result {
+        out.write_all(&[1])?;
+        serde_json::to_writer(out, &self.0)
+            .map(|_| IsNull::No)
+            .map_err(Into::into)
     }
 }